Add append procedures
This adds 2 append procedures, one which appends the items of a tensor and one that appends the items of an open array. They are otherwise the same.

These new functions are similar to (but more limited than) numpy's append. The main differences are that:
- They require their tensor inputs to be of rank-1 (i.e. they require their inputs to be explicitly flattened if their rank is higher than 1).
- They do not support the axis argument.

IMHO the automatic flattening of the output of numpy's append can lead to very surprising behavior (hence limitation #1). Only supporting rank-1 tensors makes the axis argument unnecessary (hence limitation #2).

Users that need to work with higher ranked tensors can always use the existing concat procedure.